minor changes for plots

+++ trafo/box_mutating/round_bounding_box.py
@@ -1,4 +1,4 @@
-import math
+import numpy as np
 
 from ..trafo import Trafo
 
@@ -13,4 +13,4 @@ class RoundBoundingBox(Trafo):
 def _(self, boxes: np.ndarray) -> np.ndarray:
     boxes[:, :2] = np.floor(boxes[:, :2])
     boxes[:, 2:] = np.ceil(boxes[:, 2:])
-    return boxes
+    return boxes.astype('int')
